login encaminado

+++ opercam/src/main/java/com/app/opercam/model/repository/ICredencialRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.app.opercam.model.entity.Credencial;
 
-public interface ICredencialRepository extends CrudRepository<Credencial, Integer> {
+public interface ICredencialRepository extends CrudRepository<Credencial, Long> {
 
 }
